fix the structure

+++ include/signature/area/area_hit.h
@@ -1,28 +1,28 @@
 #pragma once
 
 #include <cstdint>
-#include <singleton/action.h>
+#include <singleton/g_action.h>
 
 class Area_Hit {
     public:
-        Area_Hit(int32_t* damage, int32_t x, int32_t y, int32_t w, int32_t h);
+        Area_Hit(int32_t* damage, float x, float y, float w, float h);
 
         int32_t* get_damage();
         void set_damage(int32_t* value);
 
-        int32_t get_x();
-        void set_x(int32_t value);
+        float get_x();
+        void set_x(float value);
 
-        int32_t get_y();
-        void set_y(int32_t value);
+        float get_y();
+        void set_y(float value);
 
-        int32_t get_w();
-        void set_w(int32_t value);
+        float get_w();
+        void set_w(float value);
 
-        int32_t get_h();
-        void set_h(int32_t value);
+        float get_h();
+        void set_h(float value);
 
     private:
         int32_t* damage;
-        int32_t x, y, w, h;
+        float x, y, w, h;
 };
