+++ src/signature/area/area_hit.cpp
@@ -1,6 +1,6 @@
 #include <signature/area/area_hit.h>
 
-Area_Hit::Area_Hit(int32_t* damage, int32_t x, int32_t y, int32_t w, int32_t h){
+Area_Hit::Area_Hit(int32_t* damage, float x, float y, float w, float h){
     this->set_damage(damage);
     this->set_x(x);
     this->set_y(y);
@@ -18,34 +18,34 @@ void Area_Hit::set_damage(int32_t* value){
     this->damage = value;
 }
 
-int32_t Area_Hit::get_x(){
+float Area_Hit::get_x(){
     return this->x;
 }
 
-void Area_Hit::set_x(int32_t value){
+void Area_Hit::set_x(float value){
     this->x = value;
 }
 
-int32_t Area_Hit::get_y(){
+float Area_Hit::get_y(){
     return this->y;
 }
 
-void Area_Hit::set_y(int32_t value){
+void Area_Hit::set_y(float value){
     this->y = value;
 }
 
-int32_t Area_Hit::get_w(){
+float Area_Hit::get_w(){
     return this->w;
 }
 
-void Area_Hit::set_w(int32_t value){
+void Area_Hit::set_w(float value){
     this->w = value;
 }
 
-int32_t Area_Hit::get_h(){
+float Area_Hit::get_h(){
     return this->h;
 }
 
-void Area_Hit::set_h(int32_t value){
+void Area_Hit::set_h(float value){
     this->h = value;
 }
